CarList screen updated.

+++ app/src/main/java/com/direnaydin/arabamcom/network/CarService.kt
@@ -21,6 +21,6 @@ interface CarService {
 
     @GET(ApplicationConstants.DETAIL_API_ENDPOINT)
     suspend fun getCarDetail(
-        @Query("id") id: Long
+        @Query("id") id: Int
     ): CarDetail
 }
